Remove utils.rs; Move helpers to main.rs
